Moving the get_options function into the __init__.py instead of utils.py.
