+++ salt/returners/mongo_return.py
@@ -37,7 +37,7 @@ the default location::
 # Import python libs
 import logging
 
-import salt.returners.utils
+import salt.returners
 
 # Import third party libs
 try:
@@ -79,11 +79,11 @@ def _get_options(ret):
              'username': 'username',
              'password': 'password'}
 
-    _options = salt.returners.utils.get_returner_options(__virtualname__,
-                                                         ret,
-                                                         attrs,
-                                                         __salt__=__salt__,
-                                                         __opts__=__opts__)
+    _options = salt.returners.get_returner_options(__virtualname__,
+                                                   ret,
+                                                   attrs,
+                                                   __salt__=__salt__,
+                                                   __opts__=__opts__)
     return _options
