+++ salt/returners/redis_return.py
@@ -32,7 +32,7 @@ the default location::
 # Import python libs
 import json
 
-import salt.returners.utils
+import salt.returners
 
 # Import third party libs
 try:
@@ -59,11 +59,11 @@ def _get_options(ret=None):
              'port': 'port',
              'db': 'db'}
 
-    _options = salt.returners.utils.get_returner_options(__virtualname__,
-                                                         ret,
-                                                         attrs,
-                                                         __salt__=__salt__,
-                                                         __opts__=__opts__)
+    _options = salt.returners.get_returner_options(__virtualname__,
+                                                   ret,
+                                                   attrs,
+                                                   __salt__=__salt__,
+                                                   __opts__=__opts__)
     return _options
