+++ salt/returners/smtp_return.py
@@ -70,7 +70,7 @@ import logging
 import smtplib
 from email.utils import formatdate
 
-import salt.returners.utils
+import salt.returners
 
 try:
     import gnupg
@@ -102,11 +102,11 @@ def _get_options(ret=None):
              'fields': 'fields',
              'tls': 'tls'}
 
-    _options = salt.returners.utils.get_returner_options(__virtualname__,
-                                                         ret,
-                                                         attrs,
-                                                         __salt__=__salt__,
-                                                         __opts__=__opts__)
+    _options = salt.returners.get_returner_options(__virtualname__,
+                                                   ret,
+                                                   attrs,
+                                                   __salt__=__salt__,
+                                                   __opts__=__opts__)
     return _options
