+++ salt/returners/sqlite3_return.py
@@ -66,7 +66,7 @@ import logging
 import json
 import datetime
 
-import salt.returners.utils
+import salt.returners
 
 # Better safe than sorry here. Even though sqlite3 is included in python
 try:
@@ -94,11 +94,11 @@ def _get_options(ret=None):
     attrs = {'database': 'database',
              'timeout': 'timeout'}
 
-    _options = salt.returners.utils.get_returner_options(__virtualname__,
-                                                         ret,
-                                                         attrs,
-                                                         __salt__=__salt__,
-                                                         __opts__=__opts__)
+    _options = salt.returners.get_returner_options(__virtualname__,
+                                                   ret,
+                                                   attrs,
+                                                   __salt__=__salt__,
+                                                   __opts__=__opts__)
     return _options
